chore: Hybrid agent Client Database updates (#3157)
