+++ src/Agent/NewRelic/Agent/Core/OpenTelemetryBridge/INewRelicActivitySource.cs
@@ -1,0 +1,14 @@
+// Copyright 2020 New Relic, Inc. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+using System;
+using System.Diagnostics;
+using NewRelic.Agent.Extensions.Api.Experimental;
+
+namespace NewRelic.Agent.Core.OpenTelemetryBridge
+{
+    public interface INewRelicActivitySource : IDisposable
+    {
+        INewRelicActivity CreateActivity(string activityName, ActivityKind kind);
+    }
+}
