+++ src/Agent/NewRelic/Agent/Extensions/NewRelic.Agent.Extensions/Api/Experimental/IAgentExperimental.cs
@@ -49,6 +49,7 @@ namespace NewRelic.Agent.Api.Experimental
         Extensions.Helpers.IStackExchangeRedisCache StackExchangeRedisCache { get; set; }
 
         ISimpleSchedulingService SimpleSchedulingService { get; }
+        IDatabaseService DatabaseService { get; }
 
         void RecordLlmEvent(string eventType, IDictionary<string, object> attributes);
