+++ src/Agent/NewRelic/Agent/Extensions/NewRelic.Agent.Extensions/Api/Experimental/IDatabaseService.cs
@@ -1,0 +1,14 @@
+// Copyright 2020 New Relic, Inc. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+using System;
+using NewRelic.Agent.Extensions.Providers.Wrapper;
+
+namespace NewRelic.Agent.Api.Experimental
+{
+    public interface IDatabaseService : IDisposable
+    {
+        long GetSqlId(string sql, DatastoreVendor vendor);
+        string GetObfuscatedSql(string sql, DatastoreVendor vendor);
+    }
+}
