+++ tests/Agent/UnitTests/Core.UnitTest/CrossAgentTests/RumTests/RumClientConfigTests.cs
@@ -29,6 +29,7 @@ using NewRelic.Agent.Core.SharedInterfaces.Web;
 using NewRelic.Agent.Core.AgentHealth;
 using NewRelic.Agent.Core.Utilities;
 using NewRelic.Agent.Core.Events;
+using NewRelic.Agent.Api.Experimental;
 
 namespace NewRelic.Agent.Core.CrossAgentTests.RumTests
 {
