+++ tests/Agent/UnitTests/Core.UnitTest/Transformers/DatastoreSegmentTransformerTests.cs
@@ -14,6 +14,7 @@ using NewRelic.Agent.Extensions.Parsing;
 using NewRelic.Agent.Core.Segments;
 using NewRelic.Agent.Core.Database;
 using NewRelic.Agent.Core.Segments.Tests;
+using NewRelic.Agent.Api.Experimental;
 
 namespace NewRelic.Agent.Core.Transformers
 {
