+++ tests/Agent/UnitTests/Core.UnitTest/Wrapper/AgentWrapperApi/AgentWrapperApiTests.cs
@@ -42,6 +42,7 @@ using NUnit.Framework;
 using Telerik.JustMock;
 using NewRelic.Agent.Core.Labels;
 using NewRelic.Agent.Core.OpenTelemetryBridge;
+using NewRelic.Agent.Core.Database.UnitTest;
 
 namespace NewRelic.Agent.Core.Wrapper.AgentWrapperApi
 {
@@ -102,6 +103,7 @@ namespace NewRelic.Agent.Core.Wrapper.AgentWrapperApi
         private const string ReferrerProcessId = "referrerProcessId";
 
         private ICustomEventTransformer _customEventTransformer;
+        private IDatabaseService _databaseService;
 
         [SetUp]
         public void SetUp()
@@ -158,8 +160,9 @@ namespace NewRelic.Agent.Core.Wrapper.AgentWrapperApi
             _logContextDataFilter = new LogContextDataFilter(_configurationService);
 
             _customEventTransformer = Mock.Create<ICustomEventTransformer>();
+            _databaseService = Mock.Create<IDatabaseService>();
 
-            _agent = new Agent(_transactionService, _transactionTransformer, _threadPoolStatic, _transactionMetricNameMaker, _pathHashMaker, _catHeaderHandler, _distributedTracePayloadHandler, _syntheticsHeaderHandler, _transactionFinalizer, _browserMonitoringPrereqChecker, _browserMonitoringScriptMaker, _configurationService, _agentHealthReporter, _agentTimerService, _metricNameService, _traceMetadataFactory, _catMetrics, _logEventAggregator, _logContextDataFilter, _simpleSchedulingService, _customEventTransformer, new NewRelicActivitySourceProxy());
+            _agent = new Agent(_transactionService, _transactionTransformer, _threadPoolStatic, _transactionMetricNameMaker, _pathHashMaker, _catHeaderHandler, _distributedTracePayloadHandler, _syntheticsHeaderHandler, _transactionFinalizer, _browserMonitoringPrereqChecker, _browserMonitoringScriptMaker, _configurationService, _agentHealthReporter, _agentTimerService, _metricNameService, _traceMetadataFactory, _catMetrics, _logEventAggregator, _logContextDataFilter, _simpleSchedulingService, _customEventTransformer, new NewRelicActivitySourceProxy(), _databaseService);
         }
 
         [TearDown]
@@ -170,6 +173,7 @@ namespace NewRelic.Agent.Core.Wrapper.AgentWrapperApi
             _logContextDataFilter.Dispose();
             _attribDefSvc.Dispose();
             _labelsService.Dispose();
+            _databaseService.Dispose();
         }
 
         private class CallStackManagerFactory : ICallStackManagerFactory
@@ -675,7 +679,7 @@ namespace NewRelic.Agent.Core.Wrapper.AgentWrapperApi
         public void AcceptDistributedTraceHeaders__ReportsSupportabilityMetric_NullPayload()
         {
             _distributedTracePayloadHandler = new DistributedTracePayloadHandler(_configurationService, _agentHealthReporter, new AdaptiveSampler());
-            _agent = new Agent(_transactionService, _transactionTransformer, _threadPoolStatic, _transactionMetricNameMaker, _pathHashMaker, _catHeaderHandler, _distributedTracePayloadHandler, _syntheticsHeaderHandler, _transactionFinalizer, _browserMonitoringPrereqChecker, _browserMonitoringScriptMaker, _configurationService, _agentHealthReporter, _agentTimerService, _metricNameService, _traceMetadataFactory, _catMetrics, _logEventAggregator, _logContextDataFilter, _simpleSchedulingService, _customEventTransformer, new NewRelicActivitySourceProxy());
+            _agent = new Agent(_transactionService, _transactionTransformer, _threadPoolStatic, _transactionMetricNameMaker, _pathHashMaker, _catHeaderHandler, _distributedTracePayloadHandler, _syntheticsHeaderHandler, _transactionFinalizer, _browserMonitoringPrereqChecker, _browserMonitoringScriptMaker, _configurationService, _agentHealthReporter, _agentTimerService, _metricNameService, _traceMetadataFactory, _catMetrics, _logEventAggregator, _logContextDataFilter, _simpleSchedulingService, _customEventTransformer, new NewRelicActivitySourceProxy(), _databaseService);
             SetupTransaction();
 
             Mock.Arrange(() => _configurationService.Configuration.DistributedTracingEnabled).Returns(true);
