+++ tests/Agent/UnitTests/Core.UnitTest/Wrapper/AgentWrapperApi/CrossApplicationTracing/CatHeaderHandlerTests.cs
@@ -20,6 +20,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Telerik.JustMock;
+using NewRelic.Agent.Api.Experimental;
 
 namespace NewRelic.Agent.Core.Wrapper.AgentWrapperApi.CrossApplicationTracing
 {
